refactors and makes generic split module that split_out extends on
