+++ src/main.rs
@@ -3,6 +3,7 @@ use clap::{App, ArgMatches};
 mod commands;
 mod repo_file;
 mod split_out;
+mod split;
 mod git_helpers;
 mod exec_helpers;
